LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/tentative/static-router/static-router-resource-timing.https-expected.txt


PASS Main resource matched the rule with fetch-event source
PASS Main resource load matched with the condition and resource timing
PASS Main resource load not matched with the condition and resource timing
FAIL Main resource load matched with the cache source and resource timing assert_equals: cache as source on main resource and cache hit expected 1 but got 0
PASS Main resource fallback to the network when there is no cache entry and resource timing
FAIL Subresource load matched the rule fetch-event source assert_equals: fetch-event as source on sub resource expected "fetch-event" but got ""
FAIL Subresource load not matched with URLPattern condition assert_greater_than: no source type matched expected a number greater than 0 but got 0
PASS Subresource load matched with URLPattern condition
FAIL Subresource load matched with the cache source rule assert_equals: cache as source on subresource and cache hits expected "cache" but got ""
PASS Subresource load did not match with the cache and fallback to the network
PASS Main resource load matched the rule with race-network-and-fetch-handler source, and the fetch handler response is faster than the server response
PASS Main resource load matched the rule with race-network-and-fetch-handler source, and the server reseponse is faster than the fetch handler
FAIL Subresource load matched the rule with race-network-and-fetch-handler source, and the fetch handler response is faster than the server response assert_equals: race as source on subresource and fetch wins expected "race-network-and-fetch-handler" but got ""
PASS Subresource load matched the rule with race-network-and-fetch-handler source, and the server reseponse is faster than the fetch handler
